python-services/requirements.txt
# Core FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.3

# Security dependencies
python-jose[cryptography]>=3.3.0  # JWT authentication
passlib[bcrypt]>=1.7.4  # Password hashing
python-multipart>=0.0.9  # File uploads
slowapi>=0.1.9  # Rate limiting (alternative)

# OCR and PDF processing
pytesseract>=0.3.10
pdf2image>=1.17.0

# HTTP client
requests>=2.31.0
httpx>=0.26.0

# Vector database
qdrant-client>=1.7.1

# ML and NLP
sentence-transformers>=2.3.1
torch>=2.1.2
transformers>=4.36.2
numpy>=1.26.3

# LLM integration
openai>=1.10.0
anthropic>=0.8.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
hypothesis>=6.98.3

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
redis>=5.0.1
